refactor: LineResponse#getStations 리턴 타입 변경
List<Station> -> List<StationResponse>
