+++ src/main/java/nextstep/subway/applicaion/LineService.java
@@ -1,5 +1,7 @@
 package nextstep.subway.applicaion;
 
+import java.util.List;
+import java.util.stream.Collectors;
 import nextstep.subway.applicaion.dto.LineRequest;
 import nextstep.subway.applicaion.dto.LineResponse;
 import nextstep.subway.domain.Line;
@@ -7,9 +9,6 @@ import nextstep.subway.domain.LineRepository;
 import nextstep.subway.exception.NotFoundLineException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -44,21 +43,14 @@ public class LineService {
     }
 
     private LineResponse createLineResponse(Line line) {
-        return new LineResponse(
-            line.getId(),
-            line.getName(),
-            line.getColor(),
-            line.getStations().getStations(),
-            line.getCreatedDate(),
-            line.getModifiedDate()
-        );
+        return LineResponse.from(line);
     }
 
     public void changeLine(Long id, LineRequest lineRequest) {
         Line line = lineRepository.findById(id)
             .orElseThrow(() -> new NotFoundLineException(id));
 
-        line.update(lineRequest);
+        line.update(lineRequest.getColor(), lineRequest.getName());
     }
 
     public void deleteLine(Long id) {
