+++ src/main/java/nextstep/subway/applicaion/dto/LineResponse.java
@@ -1,20 +1,20 @@
 package nextstep.subway.applicaion.dto;
 
-import nextstep.subway.domain.Station;
-
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
+import nextstep.subway.domain.Line;
 
 public class LineResponse {
 
     private Long id;
     private String name;
     private String color;
-    private List<Station> stations;
+    private List<StationResponse> stations;
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public LineResponse(Long id, String name, String color, List<Station> stations,
+    private LineResponse(Long id, String name, String color, List<StationResponse> stations,
         LocalDateTime createdDate, LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
@@ -22,6 +22,19 @@ public class LineResponse {
         this.stations = stations;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
+    }
+
+    public static LineResponse from(Line line) {
+        return new LineResponse(
+            line.getId(),
+            line.getName(),
+            line.getColor(),
+            line.getStations().stream()
+                .map(StationResponse::from)
+                .collect(Collectors.toList()),
+            line.getCreatedDate(),
+            line.getModifiedDate()
+        );
     }
 
     public Long getId() {
@@ -36,7 +49,7 @@ public class LineResponse {
         return color;
     }
 
-    public List<Station> getStations() {
+    public List<StationResponse> getStations() {
         return stations;
     }
