+++ src/main/java/nextstep/subway/applicaion/dto/StationResponse.java
@@ -1,6 +1,7 @@
 package nextstep.subway.applicaion.dto;
 
 import java.time.LocalDateTime;
+import nextstep.subway.domain.Station;
 
 public class StationResponse {
 
@@ -9,12 +10,21 @@ public class StationResponse {
     private LocalDateTime createdDate;
     private LocalDateTime modifiedDate;
 
-    public StationResponse(Long id, String name, LocalDateTime createdDate,
+    private StationResponse(Long id, String name, LocalDateTime createdDate,
         LocalDateTime modifiedDate) {
         this.id = id;
         this.name = name;
         this.createdDate = createdDate;
         this.modifiedDate = modifiedDate;
+    }
+
+    public static StationResponse from(Station station) {
+        return new StationResponse(
+            station.getId(),
+            station.getName(),
+            station.getCreatedDate(),
+            station.getModifiedDate()
+        );
     }
 
     public Long getId() {
